+++ src/zenuml/components/MainHeader/ProductVersionInfo/ProductVersionAnonymous.js
